refactor invoice item
