+++ packages/web/app/components/ThreeDotMenu.jsx
@@ -30,7 +30,7 @@ const StyledIconButton = styled(IconButton)`
   padding: 7px;
 `;
 
-export const ThreeDotMenu = ({ items, disabled }) => {
+export const ThreeDotMenu = ({ items, disabled, className }) => {
   const [anchorEl, setAnchorEl] = useState(null);
   const open = Boolean(anchorEl);
 
@@ -52,6 +52,7 @@ export const ThreeDotMenu = ({ items, disabled }) => {
       <StyledIconButton
         onClick={onOpenKebabMenu}
         disabled={disabled}
+        className={className}
         data-testid="stylediconbutton-szh8"
       >
         <MoreVert data-testid="morevert-kusc" />
